jobs/strr-backfiller/pyproject.toml
[tool.poetry]
name = "strr-backfiller"
version = "1.0.0"
description = ""
authors = []
packages = [{include = "backfiller", from = "src"}]
license = "BSD 3-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.2"
flask = "^3.0.3"
flask-sqlalchemy = "^3.0.3"
flask-migrate = "^4.0.4"
launchdarkly-server-sdk = "^9.0.1"
python-dotenv = "^1.0.1"
sentry-sdk = "^1.44.1"
flask-babel = "^4.0.0"
gunicorn = "^21.2.0"
pg8000 = "^1.31.2"
gcp-queue = { git = "https://github.com/bcgov/sbc-connect-common.git", subdirectory = "python/gcp-queue", branch = "main" }
structured-logging = { git = "https://github.com/bcgov/sbc-connect-common.git", subdirectory = "python/structured-logging", branch = "main" }
strr-api = {git = "https://github.com/bcgov/STRR.git", rev = "main", subdirectory = "strr-api"}
nanoid = "^2.0.0"
flake8-pyproject = "^1.2.3"

[tool.poetry.group.test.dependencies]
freezegun = "^1.4.0"
hypothesis = "^6.100.0"
pexpect = "^4.9.0"
pytest = "^7.4.4"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
requests-mock = "^1.12.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
faker = "^24.4.0"
lovely-pytest-docker = "^0.3.1"
isort = "^5.13.2"
black = "^24.10.0"
pylint = "^3.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pylint]
max-line-length = 132

[tool.flake8]
ignore = ["F401","E402", "Q000", "E203", "W503"]
exclude = [
    ".venv",
    ".git",
    ".history",
    "devops",
    "*migrations*",
]
max-line-length = 132
